+++ test/amqp/vhosts/index.test.ts
@@ -19,7 +19,10 @@ const rabbitMQConfig: RabbitMQConfig = {
   port: 123,
   username: 'user',
   password: 'pass',
+  consumerPrefetch: 10,
+  producerPrefetch: 10,
 };
+
 const sandbox = sinon.createSandbox();
 let clock: any;
 
@@ -42,7 +45,12 @@ describe('RabbitMQ', () => {
     fnInit = sandbox.stub();
 
     connection = {
-      createChannel: sandbox.stub(),
+      createChannel: sandbox.stub().returns({
+        prefetch: sandbox.stub(),
+      }),
+      createConfirmChannel: sandbox.stub().returns({
+        prefetch: sandbox.stub(),
+      }),
     };
     sandbox
       .stub(amqplib, 'connect')
@@ -68,20 +76,62 @@ describe('RabbitMQ', () => {
     it('should connect, handleErros and create channel', async () => {
       const testVhost = new TestVHost('test', rabbitMQConfig);
 
-      await testVhost.init();
+      // @ts-ignore
+      await testVhost.init({});
+
       // @ts-ignore
       sandbox.assert.calledWith(amqplib.connect, connConfig);
       sandbox.assert.calledOnce(fnConnectionConfig);
       sandbox.assert.calledOnce(fnHandleOnError);
+      sandbox.assert.calledOnce(connection.createConfirmChannel);
       sandbox.assert.calledOnce(connection.createChannel);
     });
+
     it('should throw error', async () => {
       const testVhost = new TestVHost('test', rabbitMQConfig);
 
       fnConnectionConfig.throws();
-      await testVhost.init();
+
+      // @ts-ignore
+      await testVhost.init({});
 
       sandbox.assert.calledOnce(fnRecconect);
+    });
+
+    it('shoud start producers and consumers', async () => {
+      const testVhost = new TestVHost('test', rabbitMQConfig);
+
+      // @ts-ignore
+      await testVhost.init({});
+
+      sandbox.assert.calledOnce(connection.createConfirmChannel);
+      sandbox.assert.calledOnce(connection.createChannel);
+    });
+
+    it('shoud start only producers', async () => {
+      const testVhost = new TestVHost('test', rabbitMQConfig);
+      const initOptions = {
+        startConsumers: false,
+        startProducers: true,
+      };
+      // @ts-ignore
+      await testVhost.init({}, initOptions);
+
+      sandbox.assert.calledOnce(connection.createConfirmChannel);
+      sandbox.assert.notCalled(connection.createChannel);
+    });
+
+    it('shoud start only consumers', async () => {
+      const testVhost = new TestVHost('test', rabbitMQConfig);
+      const initOptions = {
+        startConsumers: true,
+        startProducers: false,
+      };
+      // @ts-ignore
+      await testVhost.init({}, initOptions);
+
+      sandbox.assert.notCalled(connection.createConfirmChannel);
+      sandbox.assert.calledOnce(connection.createChannel);
     });
   });
 
@@ -89,7 +139,7 @@ describe('RabbitMQ', () => {
     class TestVHost extends RabbitMQ {
       constructor(vHost: string, config: RabbitMQConfig) {
         super(vHost, config);
-        this.channel = {
+        this.producerChannel = {
           // @ts-ignore
           publish: (exchange, routingKey, message) => {
             fnPublish(exchange, routingKey, message);
@@ -110,13 +160,16 @@ describe('RabbitMQ', () => {
       );
     });
 
-    it('should throw error', () => {
+    it('should throw error', async () => {
       const testVhost = new TestVHost('test', rabbitMQConfig);
       fnPublish.throws();
 
-      expect(() => {
-        testVhost.send('test', 'routingKey', { message: 'test' });
-      }).to.throws();
+      try {
+        await testVhost.send('test', 'routingKey', { message: 'test' });
+        sandbox.assert.fail('should throwed error');
+      } catch (error) {
+        expect(error).to.be.instanceOf(Error);
+      }
     });
   });
 
@@ -134,7 +187,8 @@ describe('RabbitMQ', () => {
       }
 
       handleOnError() {
-        super.handleOnError();
+        // @ts-ignore
+        super.handleOnError({});
       }
 
       reconnect() {
@@ -178,11 +232,12 @@ describe('RabbitMQ', () => {
       }
 
       reconnect() {
-        super.reconnect();
+        // @ts-ignore
+        super.reconnect({});
       }
 
-      getChannel() {
-        return this.channel;
+      getConsumerChannel() {
+        return this.consumerChannel;
       }
 
       getConnection() {
@@ -193,13 +248,9 @@ describe('RabbitMQ', () => {
     it('should delete channel prop', () => {
       const testVhost = new TestVHost('test', rabbitMQConfig);
       testVhost.reconnect();
-      expect(testVhost.getChannel()).to.be.undefined;
+      expect(testVhost.getConsumerChannel()).to.be.undefined;
     });
-    it('should delete connection prop', () => {
-      const testVhost = new TestVHost('test', rabbitMQConfig);
-      testVhost.reconnect();
-      expect(testVhost.getConnection()).to.be.undefined;
-    });
+
     it('should call init method after 5000ms', () => {
       const testVhost = new TestVHost('test', rabbitMQConfig);
       testVhost.reconnect();